+++ tests/streaming/teststreaming3.cpp
@@ -263,7 +263,7 @@ int main(int argc, char *argv[])
         return -1;
     }
 
-    debugOutput(DEBUG_LEVEL_NORMAL, "verbose level = %ld\n", arguments.verbose);
+    debugOutput(DEBUG_LEVEL_NORMAL, "verbose level = %d\n", (int)arguments.verbose);
     setDebugLevel(arguments.verbose);
 
     int nb_in_channels=0, nb_out_channels=0;
